Fixing broken PGI build
Working around same PGI bug as before, caused by my rejiggering of get_int()
to use int64_t last week.  This required making sure that cstdlib was the
first file #included in a few files again, as before...



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19374 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ compiler/AST/primitive.cpp
@@ -554,7 +554,7 @@ Map<const char*, VarSymbol*> memDescsMap;
 Vec<const char*> memDescsVec;
 
 VarSymbol* newMemDesc(const char* str) {
-  static int memDescInt = 0;
+  static int64_t memDescInt = 0;
   const char* s = astr(str);
   if (VarSymbol* v = memDescsMap.get(s))
     return v;
